+++ src/main/java/com/tp7/flutter/repos/ClasseRepository.java
@@ -3,4 +3,4 @@ package com.tp7.flutter.repos;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.tp7.flutter.entities.Classe;
 
-public interface ClasseRepository extends JpaRepository<Classe, Integer> {}
+public interface ClasseRepository extends JpaRepository<Classe, Long> {}
