+++ src/main/java/com/tp7/flutter/repos/EtudiantRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import com.tp7.flutter.entities.Etudiant;
 
-public interface EtudiantRepository extends JpaRepository<Etudiant, Integer> {
-    List<Etudiant> findByClasse_Id(Integer classeId);
+public interface EtudiantRepository extends JpaRepository<Etudiant, Long> {
+    List<Etudiant> findByClasse_Id(Long classeId);
 }
