+++ src/main/java/com/tp7/flutter/repos/FormationRepository.java
@@ -3,4 +3,4 @@ package com.tp7.flutter.repos;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.tp7.flutter.entities.Formation;
 
-public interface FormationRepository extends JpaRepository<Formation, Integer> {}
+public interface FormationRepository extends JpaRepository<Formation, Long> {}
