+++ src/main/java/com/tp7/flutter/repos/MatiereRepository.java
@@ -3,4 +3,4 @@ package com.tp7.flutter.repos;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.tp7.flutter.entities.Matiere;
 
-public interface MatiereRepository extends JpaRepository<Matiere, Integer> {}
+public interface MatiereRepository extends JpaRepository<Matiere, Long> {}
